Do not fail in ipmi-configure if calls to ipmitool fail
I guess it's buggy hardware, but we're seeing calls to
    ipmitool lan set 1 defgw ipaddr
fail on some hardware. Unfortunately, this happens on
hardware-installing and this means that everything after ipmi-configure
will not run. Including the bootdisk finder in provisioner.

So instead of using a bash block, we use a ruby_block where we can see
the exit value of ipmitool.