Allow the admin node for ipmi-configure/discover
By default it seems the admin node is included in the
proposal, but when you try to save/apply it, you get
the error message that the proposal cannot be saved/applied
due to the role not being valid for admin. Since the
admin node has a central role here it seems, we should
allow it.

(cherry picked from commit 68d2f875c5b41e7b619ab207ee558a90957d50f9)