Add optional crowbar-applied boolean attribute to schema
We add a crowbar applied attribute to mark the proposal as applied after the
chef run. This allows the UI to show that the current form of the proposal has
been 'applied' successfully (i.e. chef-client completed the run w/ success) and
that the proposal was then not modified since that.

Update the proposal_schema_revision, so we can add the attribute to schema for
('3rd party') barclamps that currently do not have it.

Refs: crowbar/crowbar#2044 (which adds this attribute to barclamps which lack
it), crowbar/barclamp-crowbar#1068 (which marks the proposals as applied) and
https://bugzilla.novell.com/show_bug.cgi?id=877486